Extract definition lookup service class
