+++ script/build_cards.rb
@@ -1,30 +1,6 @@
 # frozen_string_literal: true
 
-require 'dotenv/load'
-require 'http'
-require 'cgi'
-
-# Wrap M-W Dictionary API, providing a simple interface for fetching word definitions
-class DefinitionLookupService
-  attr_reader :key, :url
-
-  def initialize
-    @url = 'https://dictionaryapi.com/api/v3/references/collegiate/json/'
-    @key = ENV['DICTIONARY_API_KEY']
-  end
-
-  def parse_response(json)
-    JSON.parse(json)
-    # TODO: extract the definition(s) from the response
-  end
-
-  def definitions_for(word)
-    response = HTTP.get(url + CGI.escape(word) + "?key=#{key}")
-    return false unless response.status.success?
-
-    parse_response(response.body)
-  end
-end
+require_relative '../lib/definition_lookup_service'
 
 # read in all spellings words from a file
 def gather_words(input_file)
